fix Media.from_result
